Optimizing discount calculations and adding scenario tests

+++ ShoppingBasket.Core/Models/Product.cs
@@ -1,9 +1,5 @@
 ﻿using ShoppingBasket.Core.Enumerations;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ShoppingBasket.Models.Core
 {
@@ -11,7 +7,7 @@ namespace ShoppingBasket.Models.Core
 	{
 		public Guid Id { get; set; }
 		public string Name { get; set; }
-		public double Price { get; set; }
+		public decimal Price { get; set; }
 		public bool IsDiscountApplied { get; set; }
 		public ProductType Type { get; set; }
 	}
